requirements.txt
# Flask and Extensions
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41

# Third-party APIs
twilio==9.6.1
openai==1.3.8
stripe==6.7.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
werkzeug==2.3.7
pyjwt==2.8.0
bcrypt==4.0.1
requests==2.31.0
phonenumbers==8.13.7

# Security
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-cov==4.1.0

pytz

